fixed label v1
fix: Interaction was possible from everywhere on the map because of collision layer of campfire set on default (1) now is set to 2, 3 (mask to 3 - important for interaction, player has now collision layer 3 as well). Label now works as intended, collision as well, interaction too.

Ref..: "I realized it's because while my InteractionArea's collisionshape was masked to only detect the player, the actual object it was attached to was set on the default layer, which is player.  It fired immediately on startup because it thought it was the player lmao."